old/logic/requirements.txt
whisper
moviepy
scenedetect
fuzzywuzzy
python-levenshtein
minio 
psycopg2-binary
sqlalchemy 
elasticsearch
fastapi 
uvicorn 
ffmpeg-python
